Clean up whatever that was
I love working on the same project on two computers and *not* syncing changes :D
