+++ JSharp.Example/Expected.cs
@@ -1,15 +1,6 @@
-using JSharp.Attributes;
-using JSharp.Attributes.Processors;
-
-namespace JSharp.Example;
+﻿namespace JSharp.Example;
 
 public class Expected
 {
-    public static void Main(string[] args)
-    {
-        foreach (var line in JClassProcessor.GenerateBytecode(typeof(Expected)))
-        {
-            Console.WriteLine(BitConverter.ToString(line).Replace("-", " "));
-        }
-    }
+
 }
