--- JSharp/Attributes/JClassAttribute.cs
@@ -1,7 +1,0 @@
-namespace JSharp.Attributes;
-
-[AttributeUsage(AttributeTargets.Class)]
-public class JClassAttribute : Attribute
-{
-
-}
