--- JSharp/Class/Pool/ConstantPoolType.cs
@@ -1,22 +1,0 @@
-namespace JSharp.Class.Pool;
-
-internal enum ConstantPoolType
-{
-    String = 1,
-    Integer = 3,
-    Float = 4,
-    Long = 5,
-    Double = 6,
-    Class = 7,
-    StringRef = 8,
-    FieldRef = 9,
-    MethodRef = 10,
-    InterfaceMethodRef = 11,
-    NameAndType = 12,
-    MethodHandle = 15,
-    MethodType = 16,
-    Dynamic = 17,
-    InvokeDynamic = 18,
-    Module = 19,
-    Package = 20
-}
