+++ JSharp/JSharp.cs
@@ -8,77 +8,78 @@ namespace JSharp;
 
 public class JSharp : Task
 {
-    private static readonly HashSet<Assembly> Assemblies = new();
-    
+    private readonly HashSet<Assembly> _assemblies = new();
+
     // As these properties get filled in by the .csproj calling this task,
     // the CS8618 warning actually doesn't apply and can be suppressed.
 #pragma warning disable CS8618
-    [Required]
-    public string AssemblyLocation { get; set; }
+    [Required] public string AssemblyLocation { get; set; }
 #pragma warning restore CS8618
-    
+
     public override bool Execute()
     {
         // Compile list of assemblies
         var locations = AssemblyLocation.Split(';');
         foreach (var asm in locations)
-        {
             try
             {
-                Assemblies.Add(Assembly.LoadFile(asm));
+                Log.LogWarning($"Provided assembly '{asm}'");
+                var asmBytes = File.ReadAllBytes(asm);
+                _assemblies.Add(Assembly.Load(asmBytes));
             }
             catch (ArgumentException argumentException)
             {
                 if (argumentException is ArgumentNullException)
                 {
-                    Log.LogError(subcategory: null,
-                        errorCode: "JSHRP0001",
-                        helpKeyword: null,
-                        file: null,
-                        lineNumber: 0,
-                        columnNumber: 0,
-                        endColumnNumber: 0,
-                        endLineNumber: 0,
-                        message: ErrorCodes.JSHRP0001);
+                    Log.LogError(null,
+                                 "JSHRP0001",
+                                 null,
+                                 null,
+                                 0,
+                                 0,
+                                 endColumnNumber: 0,
+                                 endLineNumber: 0,
+                                 message: ErrorCodes.JSHRP0001);
                     return false;
                 }
-                Log.LogError(subcategory: null,
-                    errorCode: "JSHRP0002",
-                    helpKeyword: null,
-                    file: null,
-                    lineNumber: 0,
-                    columnNumber: 0,
-                    endColumnNumber: 0,
-                    endLineNumber: 0,
-                    message: ErrorCodes.JSHRP0002);
+
+                Log.LogError(null,
+                             "JSHRP0002",
+                             null,
+                             null,
+                             0,
+                             0,
+                             endColumnNumber: 0,
+                             endLineNumber: 0,
+                             message: ErrorCodes.JSHRP0002);
                 return false;
             }
             catch (FileNotFoundException notFoundException)
             {
-                Log.LogError(subcategory: null,
-                    errorCode: "JSHRP0003",
-                    helpKeyword: null,
-                    file: null,
-                    lineNumber: 0,
-                    columnNumber: 0,
-                    endColumnNumber: 0,
-                    endLineNumber: 0,
-                    message: ErrorCodes.JSHRP0003,
-                    messageArgs: AssemblyLocation);
+                Log.LogError(null,
+                             "JSHRP0003",
+                             null,
+                             null,
+                             0,
+                             0,
+                             endColumnNumber: 0,
+                             endLineNumber: 0,
+                             message: ErrorCodes.JSHRP0003,
+                             messageArgs: AssemblyLocation);
                 return false;
             }
-        }
 
         foreach (var type
-                 in from assembly 
-                     in Assemblies from type 
-                     in assembly.DefinedTypes 
-                 where type.CustomAttributes.Any(attr => attr.AttributeType != typeof(JExcludeAttribute))
-                 select type)
+                 in from assembly
+                        in _assemblies
+                    from type
+                        in assembly.DefinedTypes
+                    where type.CustomAttributes.Any(attr => attr.AttributeType != typeof(JExcludeAttribute))
+                    select type)
         {
-            var classBytes = new JClassProcessor(type).ClassBytes;
+            var classBytes = new JClassProcessor(type).GenerateBytecode();
         }
-        
+
         return true;
     }
 }
