+++ JSharp/Metadata/AttributeInfo.cs
