+++ JSharp/Metadata/CodeAttributeInfo.cs
