+++ JSharp/Pool/ConstantPoolValue.cs
@@ -1,28 +1,17 @@
-﻿namespace JSharp.Pool;
+namespace JSharp.Pool;
 
-public struct ConstantPoolValue
+internal struct ConstantPoolValue
 {
-    public ConstantPoolType Type;
-    public object Value;
-    public int Index;
+    public ConstantPoolType Type { get; init; }
+    public object Value { get; init; }
+    public int Index { get; set; }
+    public int OwningIndex { get; set; }
+
+    private bool Equals(ConstantPoolValue other) => Type == other.Type && Value.Equals(other.Value);
+
+    public override bool Equals(object? obj) => obj is ConstantPoolValue other && Equals(other);
 
     public static explicit operator ConstantPoolValue(string s) => new() {Type = ConstantPoolType.String, Value = s};
 
-    public bool Equals(ConstantPoolValue other)
-    {
-        return Type == other.Type && Value.Equals(other.Value);
-    }
-
-    public override bool Equals(object? obj)
-    {
-        return obj is ConstantPoolValue other && Equals(other);
-    }
-
-    public override int GetHashCode()
-    {
-        unchecked
-        {
-            return ((int) Type * 397) ^ Value.GetHashCode();
-        }
-    }
+    public override int GetHashCode() => HashCode.Combine((int) Type, Value);
 }
