+++ JSharp/System/Runtime/CompilerServices/IsExternalInit.cs
@@ -1,0 +1,7 @@
+﻿// ReSharper disable once CheckNamespace
+namespace System.Runtime.CompilerServices;
+
+internal class IsExternalInit
+{
+
+}
